requirements.txt
aiohttp==3.7.4.post0
astroid==2.5.3
async-timeout==3.0.1
attrs==21.2.0
beautifulsoup4==4.9.3
chardet==4.0.0
discord.py==1.7.3
idna==3.2
isort==5.8.0
lazy-object-proxy==1.6.0
MarkupSafe==1.1.1
mccabe==0.6.1
multidict==5.1.0
py-iMessage==1.7
pylint==2.7.4
python-decouple==3.4
python-dotenv==0.19.0
pytz==2019.3
selenium==3.141.0
soupsieve==2.2.1
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.6
wrapt==1.12.1
yarl==1.6.3
